app/backend/requirements.txt
phenex==0.7.4
fastapi[standard]
gunicorn
uvicorn
openai
httpx
typing-extensions>=4.11.0,<5.0.0
pydantic-ai[openai]
deepdiff
fastembed
faiss-cpu
python-jose[cryptography]
msal
passlib[bcrypt]
python-multipart
asyncpg
confuse
authlib
SQLAlchemy
psycopg[binary]
argon2-cffi
medconb-client
